Moved cash payment related things
