+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/cashpayment/WooPosCashPaymentState.kt
@@ -1,0 +1,8 @@
+package com.woocommerce.android.ui.woopos.cashpayment
+
+data class WooPosCashPaymentState(
+    val enteredAmount: String,
+    val changeDue: String,
+    val total: String,
+    val canBeOrderBeCompleted: Boolean,
+)
